+++ backlib/internal/backports/py311/errno/internal/darwin1.py
@@ -1,0 +1,4 @@
+from typing import Final
+
+
+EQFULL: Final[int] = 106