+++ backlib/internal/backports/py311/errno/internal/freebsd13.py
@@ -1,0 +1,4 @@
+from typing import Final
+
+
+ENOTCAPABLE: Final[int] = 93