+++ backlib/internal/backports/py312/errno.py
@@ -1,4 +1,4 @@
-from backlib.py313.internal.backports.errno.internal.errno import (
+from backlib.internal.backports.py311.errno import (
     E2BIG,
     EACCES,
     EADDRINUSE,