+++ backlib/internal/backports/py313/errno.py
@@ -1,4 +1,4 @@
-from backlib.internal.backports.py313.errno import (
+from backlib.internal.backports.py312.errno import (
     E2BIG,
     EACCES,
     EADDRINUSE,