+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Added `backlib.py310.errno`;
 * Added `backlib.py310.json`;
 * Added `backlib.py310.operator`;
+* Added `backlib.py311.errno`;
 * Added `backlib.py311.json`;
 * Added `backlib.py311.operator`;
 * Added `backlib.py311.tomllib`;
+* Added `backlib.py312.errno`;
 * Added `backlib.py312.json`;
 * Added `backlib.py312.operator`;
 * Added `backlib.py312.tomllib`;